+++ crates/common/src/cache/mod.rs
@@ -76,8 +76,8 @@ use nautilus_model::{
     },
     events::{AccountState, OrderEventAny},
     identifiers::{
-        AccountId, ClientId, ClientOrderId, ComponentId, ExecAlgorithmId, InstrumentId,
-        OrderListId, PositionId, StrategyId, Venue, VenueOrderId,
+        AccountId, ActorId, ClientId, ClientOrderId, ExecAlgorithmId, InstrumentId, OrderListId,
+        PositionId, StrategyId, Venue, VenueOrderId,
     },
     instruments::{Instrument, InstrumentAny, SyntheticInstrument},
     orderbook::{
@@ -379,7 +379,7 @@ impl<'a> CacheApi<'a> {
     ///
     /// Panics if the cache is already mutably borrowed.
     #[must_use]
-    pub fn actor_ids(&self) -> AHashSet<ComponentId> {
+    pub fn actor_ids(&self) -> AHashSet<ActorId> {
         self.cache().actor_ids()
     }
 
@@ -2750,11 +2750,11 @@ impl Cache {
     /// Returns an error if loading actor state fails.
     pub fn load_actor_state(
         &self,
-        component_id: &ComponentId,
+        actor_id: &ActorId,
     ) -> anyhow::Result<Option<IndexMap<String, Vec<u8>>>> {
         self.database
             .as_ref()
-            .map(|database| database.load_actor(component_id))
+            .map(|database| database.load_actor(actor_id))
             .transpose()
             .map(|state| state.map(Self::decode_component_state))
     }
@@ -2784,11 +2784,11 @@ impl Cache {
     /// Returns an error if updating actor state fails.
     pub fn update_actor_state(
         &self,
-        component_id: &ComponentId,
+        actor_id: &ActorId,
         state: &IndexMap<String, Vec<u8>>,
     ) -> anyhow::Result<()> {
         if let Some(database) = &self.database {
-            database.update_actor(component_id, &Self::encode_component_state(state))?;
+            database.update_actor(actor_id, &Self::encode_component_state(state))?;
         }
         Ok(())
     }
@@ -6059,9 +6059,9 @@ impl Cache {
         )
     }
 
-    /// Returns the `ComponentId`s of all actors.
+    /// Returns the `ActorId`s of all actors.
     #[must_use]
-    pub fn actor_ids(&self) -> AHashSet<ComponentId> {
+    pub fn actor_ids(&self) -> AHashSet<ActorId> {
         self.index.actors.clone()
     }