+++ crates/infrastructure/tests/test_cache_postgres.rs
@@ -42,7 +42,7 @@ mod serial_tests {
             order::spec::{OrderCancelRejectedSpec, OrderModifyRejectedSpec},
         },
         identifiers::{
-            AccountId, ClientId, ClientOrderId, ComponentId, InstrumentId, PositionId, StrategyId,
+            AccountId, ActorId, ClientId, ClientOrderId, InstrumentId, PositionId, StrategyId,
             TradeId, TraderId, VenueOrderId,
         },
         instruments::{
@@ -358,10 +358,10 @@ mod serial_tests {
         let synthetic_result = database
             .load_synthetic(&InstrumentId::from("SYNTHETIC.SYNTH"))
             .await;
-        let actor_result = database.load_actor(&ComponentId::from("ACTOR-001"));
+        let actor_result = database.load_actor(&ActorId::from("ACTOR-001"));
         let strategy_result = database.load_strategy(&StrategyId::from("STRATEGY-001"));
         let state = AHashMap::from([("state".to_string(), Bytes::from_static(b"value"))]);
-        let actor_update_result = database.update_actor(&ComponentId::from("ACTOR-001"), &state);
+        let actor_update_result = database.update_actor(&ActorId::from("ACTOR-001"), &state);
         let strategy_update_result =
             database.update_strategy(&StrategyId::from("STRATEGY-001"), &state);